Implement SSR for ColorBox
