+++ testing/tests/DevExpress.serverSide/colorBox.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets.editors/colorBox.markup.tests.js");
