+++ testing/tests/DevExpress.serverSide/colorView.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets.editors/colorView.markup.tests.js");
