+++ testing/tests/DevExpress.ui.widgets.editors/colorBox.tests.js
@@ -97,24 +97,6 @@ QUnit.module("Color Box", {
     }
 });
 
-QUnit.test("Apply css class", function(assert) {
-    var $colorBox = showColorBox.call(this);
-    assert.ok($colorBox.hasClass(COLOR_BOX_CLASS));
-});
-
-QUnit.test("Render color input", function(assert) {
-    var $colorBox = showColorBox.call(this),
-        $input = $colorBox.find("." + COLOR_BOX_INPUT_CLASS);
-
-    assert.equal($input.length, 1);
-    assert.ok($input.closest("." + COLOR_BOX_INPUT_CONTAINER_CLASS).length);
-});
-
-QUnit.test("Default value should be 'null'", function(assert) {
-    var colorBox = showColorBox.call(this).dxColorBox("instance");
-    assert.strictEqual(colorBox.option("value"), null);
-});
-
 QUnit.test("Render with hex value", function(assert) {
     var $colorBox = showColorBox.call(this, { value: "#000000" }),
         $input = $colorBox.find("." + COLOR_BOX_INPUT_CLASS);
@@ -122,28 +104,6 @@ QUnit.test("Render with hex value", function(assert) {
     assert.equal($input.val(), "#000000");
     $colorBox.dxColorBox("instance").option("value", "#ff0000");
     assert.equal($input.val(), "#ff0000");
-});
-
-QUnit.test("Render in rgba mode", function(assert) {
-    var $colorBox = showColorBox.call(this, { value: "#ff0000", editAlphaChannel: true }),
-        $input = $colorBox.find("." + COLOR_BOX_INPUT_CLASS);
-
-    assert.equal($input.val(), "rgba(255, 0, 0, 1)");
-});
-
-QUnit.test("Render color result preview", function(assert) {
-    var $colorBox = this.element.dxColorBox(),
-        $colorPreview = $colorBox.find("." + COLOR_BOX_COLOR_RESULT_PREVIEW_CLASS);
-
-    assert.equal($colorPreview.length, 1);
-    assert.equal(getColorBoxOverlay().length, 0);
-});
-
-QUnit.test("If default value is 'null' color result preview should have a special css class", function(assert) {
-    var $colorBox = this.element.dxColorBox(),
-        $colorInputContainer = $colorBox.find("." + COLOR_BOX_INPUT_CONTAINER_CLASS);
-
-    assert.ok($colorInputContainer.hasClass(COLOR_BOX_COLOR_IS_NOT_DEFINED));
 });
 
 QUnit.test("If value is set as 'null' color result preview should not have background color - the first case(T198625)", function(assert) {
@@ -154,7 +114,6 @@ QUnit.test("If value is set as 'null' color result preview should not have backg
 
     assert.ok(!colorBox.$element().find(".dx-colorbox-color-result-preview").attr("style"));
 });
-
 
 QUnit.test("If value is set as 'null' color result preview should not have background color - the second case(T198625)", function(assert) {
 
@@ -790,12 +749,4 @@ QUnit.test("Value should not be changed by 'up' key when colorbox was opened and
     keyboard.keyDown("enter");
 
     assert.equal(colorBox.option("value"), "#326b8a");
-});
-
-QUnit.test("colorbox should work with deferRendering: false", function(assert) {
-    assert.expect(0);
-
-    $("#color-box").dxColorBox({
-        deferRendering: false
-    });
 });
